fix(sqlite): bigint datatype is not supported

+++ packages/core/test/unit/query-generator/select-query.test.ts
@@ -37,7 +37,7 @@ describe('QueryGenerator#selectQuery', () => {
         autoIncrement: true,
         primaryKey: true,
       },
-      duration: DataTypes.BIGINT,
+      duration: DataTypes.INTEGER,
     }, { timestamps: false });
 
     const ProjectContributor = sequelize.define('ProjectContributor', {}, { timestamps: false });
